refactor: 유효성 검증 부분 관심사 분리하여 새로운 클래스에 정의
